Selection manager: fixes initial value
When the modal initially opens, the `#multiple` is `false`,
(regardless of how it is configured), so the initial value is set to
`[undefined]`. Then when the `#multiple` is observed as `true`
(there must be an underlying bug with the modal context code here),
then the `#selection` array already has an initial value of
`[undefined]` so will append newly selected values to that array.

I discovered this issue due to a bug with the Tree Picker editor.