Property Editor: Number - set numeric value
This was setting the value as a `string`, which would cause issues
when used in certain scenarios, e.g. MNTP min/max configuration,
the value would assumed to be numeric, but it'd be a string, so
conditions like `max === 1` would be false, as it'd be `"1"` instead.

Ensuring the value is a `number` before the "property-value-change"
event is sent appears to resolve this issue.