Tree Element: Set selection
to be done in `connectedCallback` instead of within the property
setter methods.

The issue was that logic is happening in the SelectionManager's
`setSelection()` method that relies on the `#multiple` property;
which hadn't been assigned to at this point. So the logic became faulty.

By moving the SelectionManager method calls to the `connectedCallback`,
it ensures that all of the attribute/property values have been set.