Bugfix: Re-introduces Integer, removes Number editor
In v13, the alias for the Numeric property-editor was `Umbraco.Integer`,
which in the v14 data-type migration was aligned with
`Umb.PropertyEditorUi.Integer`.

Also for v14, a `Umb.PropertyEditorUi.Number` had also been introduced,
with the Integer being a shared base for both Number and Decimal
editors. However due to a subsequent issue between the Number and
Decimal editor configurations, the Integer editor was removed.
This causes an issue for the v13 to v14 data-type migration.

This PR re-introduces the Integer editor and removes the Number
editor, and updates all the internal references.